refactor(cli): split tsserver comment editing handlers
